+++ fields/types/html/editor/constants/layout-style.js
@@ -1,0 +1,24 @@
+let mobileStyle = {
+	width: '100%',
+};
+let tabletFullStyle = {
+	maxWidth: '100%',
+};
+let tabletMaxStyle = {
+	maxWidth: '644px',
+};
+let tabletMinStyle = {
+	width: '532px',
+};
+/*
+// wait for spec
+let descktopStyle = {
+};
+*/
+
+export {
+	mobileStyle,
+	tabletMinStyle,
+	tabletMaxStyle,
+	tabletFullStyle,
+};
